move incident type options logic out of App.js
